+++ packages/docs/src/pages/form/samples/Form9.js
@@ -1,35 +1,26 @@
-import React, {useState} from 'react';
-import {
-  Button,
-  Card,
-  Checkbox,
-  Col,
-  Form,
-  Input,
-  Notification,
-  Row,
-  Select,
-  RadioGroup,
-  Radio,
-} from 'react-windy-ui';
+import React from 'react';
+import {Button, Card, Checkbox, Col, Form, Input, Notification, Radio, RadioGroup, Row, Select,} from 'react-windy-ui';
 
 export default function Form9() {
 
-  const {form, watch} = Form.useForm({
+  const {form} = Form.useForm({
     mode: 'onSubmit',
     shouldFocusError: false,
     defaultValues: {
+      firstName: 'name',
+      city: 'ShangHai',
       accept: true,
     },
   });
 
   const onSubmit = (data, e) => {
-    Notification.mini({
+    Notification.info({
       position: 'topCenter',
-      title: 'The form data:',
+      // title: 'The form data:',
       body: `${JSON.stringify(data)}`,
     });
     //then call the api to save the data
+    //.....
   };
 
   const gutter = {x: 8};
@@ -37,7 +28,6 @@ export default function Form9() {
     xs: 12,
     sm: 6,
   };
-
   return <>
     <Card style={{minWidth: '80%', padding: '.5rem'}} hasBox={true}>
       <Card.Row>
@@ -67,10 +57,8 @@ export default function Form9() {
                 <Input block/>
               </Form.Item>
             </Col>
-          </Row>
 
-          <Row gutter={gutter}>
-            <Col>
+            <Col xs={12}>
               <Form.Item
                 direction="vertical"
                 label="Address"
@@ -82,9 +70,7 @@ export default function Form9() {
                 <Input block/>
               </Form.Item>
             </Col>
-          </Row>
 
-          <Row gutter={gutter}>
             <Col xs={12} sm={4}>
               <Form.Item label="City"
                          name="city"
@@ -93,7 +79,6 @@ export default function Form9() {
                            required: 'The city is required',
                          }}>
                 <Select block placeholder="Please select a city"
-                        defaultValue=""
                         onSelect={(value) => console.log(value)}>
                   <Select.Option value="BeiJing">BeiJing</Select.Option>
                   <Select.Option value="ShangHai">ShangHai</Select.Option>
@@ -110,7 +95,6 @@ export default function Form9() {
                            required: 'The state is required',
                          }}>
                 <Select block placeholder="Please select a state"
-                        defaultValue=""
                         onSelect={(value) => console.log(value)}>
                   <Select.Option value="...">...</Select.Option>
                   <Select.Option value="other">Other</Select.Option>
@@ -125,12 +109,10 @@ export default function Form9() {
                 rules={{
                   required: 'The zip is required',
                 }}>
-                <Input placeholder="Zip" defaultValue=""/>
+                <Input placeholder="Zip"/>
               </Form.Item>
             </Col>
-          </Row>
 
-          <Row gutter={gutter}>
             <Col {...colConf}>
               <Form.Item
                 direction="vertical"
@@ -154,26 +136,25 @@ export default function Form9() {
                 <Input block/>
               </Form.Item>
             </Col>
+
+            <Form.Item name="accept" rules={{
+              validate: accepted => {
+                if (!accepted) {
+                  return 'You must accept the terms';
+                }
+              },
+            }}>
+              <Checkbox>Accept the terms</Checkbox>
+            </Form.Item>
+
+            <Form.Item compact={true}>
+              <Row>
+                <Col>
+                  <Button color="blue" nativeType="submit">Save</Button>
+                </Col>
+              </Row>
+            </Form.Item>
           </Row>
-
-          <Form.Item name="accept"
-                     rules={{
-                       validate: value => {
-                         if (!value) {
-                           return 'You must accept the terms';
-                         }
-                       },
-                     }}>
-            <Checkbox defaultChecked={false}>Accept the terms</Checkbox>
-          </Form.Item>
-
-          <Form.Item compact={true}>
-            <Row>
-              <Col>
-                <Button color="blue" nativeType="submit">Save</Button>
-              </Col>
-            </Row>
-          </Form.Item>
         </Form>
       </Card.Row>
     </Card>