+++ packages/components/CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.6.0
 * react-hook-form: upgrade to v7
+* Notification: adapt the body's width
 
 ### 0.5.1 - 0.5.8
 fix bugs for new project structure